chore(tests): move CountdownPlugin to plugin_collection
