+++ falco_plugin_tests/src/plugin_collection/source/mod.rs
@@ -1,1 +1,2 @@
 pub mod batched_empty_event;
+pub mod countdown;
